Load the WebMcpBridge runtime from a lazy webmcp-runtime chunk
The 330-line bridge class moves to webmcp-runtime-entry.ts, constructed by client.ts from a sibling chunk (2 kB gzip) only when config.webmcp.enabled is true (-1.3 kB off the IIFE). Core-owned state crosses the boundary by injection (WebMcpBridgeDeps: title map recorder/reader + slot-aware polyfill loader), so the noExternal chunk carries neither a duplicated title map nor the inlined polyfill. npm keeps the historical one-argument WebMcpBridge constructor via the webmcp-bridge-public facade, now exported from index.ts only (CDN global drops the class, following the dev-helpers precedent). executeWebMcpToolCall routes on config rather than bridge presence so an in-flight chunk can't misroute to the legacy resume path. Browser-verified: dispatch loads webmcp-runtime.js, the injected loader fetches webmcp-polyfill.js, document.modelContext installs.
