+++ packages/widget/tsconfig.json
@@ -20,7 +20,8 @@
       "@runtypelabs/persona/event-stream-view": ["src/event-stream-view.ts"],
       "@runtypelabs/persona/animations-extra": ["src/animations-extra.ts"],
       "@runtypelabs/persona/approval-ui": ["src/approval-ui.ts"],
-      "@runtypelabs/persona/session-reconnect": ["src/session-reconnect.ts"]
+      "@runtypelabs/persona/session-reconnect": ["src/session-reconnect.ts"],
+      "@runtypelabs/persona/webmcp-runtime": ["src/webmcp-runtime.ts"]
     },
     "declaration": true,
     "outDir": "dist"
