+++ packages/widget/tsup.global.config.ts
@@ -87,5 +87,10 @@ export default defineConfig({
     // scheme: sibling-URL chunk `session-reconnect.js`, fetched only when a
     // session with a `reconnectStream` transport first needs to resume.
     options.external.push("@runtypelabs/persona/session-reconnect");
+
+    // Keep the WebMCP bridge runtime (the WebMcpBridge class) out of the CDN
+    // payload. Same scheme: sibling-URL chunk `webmcp-runtime.js`, fetched by
+    // client.ts only when `config.webmcp.enabled === true`.
+    options.external.push("@runtypelabs/persona/webmcp-runtime");
   },
 });
